+++ EliteAPI/Events/Startup/GameModeType.cs
@@ -1,0 +1,11 @@
+namespace EliteAPI.Events {
+    /// <summary>
+    /// Game mode a commander plays in.
+    /// </summary>
+    public enum GameModeType
+    {
+        Open,
+        Solo,
+        Group
+    }
+}
