+++ EliteAPI/Events/Startup/MaterialsInfo.cs
@@ -36,31 +36,4 @@ namespace EliteAPI.Events.Startup
             return api.Events.InvokeMaterialsEvent(JsonConvert.DeserializeObject<MaterialsInfo>(json, JsonSettings.Settings));
         }
     }
-
-    /// <summary>
-    /// Contains information about grouped material.
-    /// </summary>
-    /// <see cref="MaterialsInfo"/>
-    public class MaterialInfo
-    {
-        internal MaterialInfo() { }
-
-        /// <summary>
-        /// The name of the material.
-        /// </summary>
-        [JsonProperty("Name")]
-        public string Name { get; internal set; }
-
-        /// <summary>
-        /// The localised name of the material.
-        /// </summary>
-        [JsonProperty("Name_Localised")]
-        public string NameLocalised { get; internal set; }
-
-        /// <summary>
-        /// The amount of materials in this group.
-        /// </summary>
-        [JsonProperty("Count")]
-        public long Count { get; internal set; }
-    }
 }
