+++ EliteAPI/Events/Startup/ShipType.cs
@@ -1,0 +1,6 @@
+namespace EliteAPI.Events.Startup {
+    public enum ShipType
+    {
+        //TODO: Get list of all ships
+    }
+}
