+++ EliteAPI/Events/Startup/VesselType.cs
@@ -1,0 +1,7 @@
+namespace EliteAPI.Events.Startup {
+    /// <summary>
+    /// The type of vessel, either 'Ship' or 'SRV'.
+    /// </summary>
+    /// <see cref="CargoInfo"/>
+    public enum VesselType { Ship, SRV }
+}
